# shared/openapi.yaml
openapi: 3.0.3
info:
  title: EasySSH API
  description: |
    EasySSH 后端 API 规范 - 现代化的 SSH 管理平台

    ## 功能模块
    - 用户认证与授权
    - 服务器管理（CRUD）
    - SSH 连接与终端
    - SFTP 文件传输
    - 批量操作与脚本执行
    - 系统监控与日志

  version: 1.0.0
  contact:
    name: EasySSH Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: 当前 EasySSH 实例
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Ory Fosite 签发的 OAuth 2.0 Access Token
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
          example: Invalid request parameters
        details:
          type: object
          additionalProperties: true
        detail:
          type: string
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    OAuthAuthorizeResponse:
      type: object
      properties:
        code:
          type: string
          description: 授权码（未启用 2FA 时返回）
          example: AbCdEfGh123
        state:
          type: string
          description: 从请求中透传返回的 state
        requires_2fa:
          type: boolean
          description: 是否需要继续进行 2FA 验证
          example: true
        temp_token:
          type: string
          description: 启用 2FA 时返回的临时令牌
...
    RoleListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        total:
          type: integer
...
paths:
...
  /oauth/authorize:
    get:
      operationId: beginOAuthAuthorization
      tags: [auth]
      summary: 标准 OAuth 2.0 授权入口
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: nonce
          in: query
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
      responses:
        '302':
          description: 跳转到 Provider 登录页面
    post:
      operationId: submitOAuthAuthorization
      tags:
        - auth
      summary: 使用邮箱密码 + PKCE 创建授权码
      description: |
        前端以 JSON 方式提交邮箱密码与 PKCE 参数，成功时返回授权码；
        如用户启用 2FA，则返回 requires_2fa=true 与临时令牌。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthAuthorizeRequest'
      responses:
        '200':
          description: 授权码创建成功或需要 2FA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthAuthorizeResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 认证失败（用户名或密码错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /oauth/logout:
    post:
      operationId: logoutOAuthSession
      tags:
        - auth
      summary: 用户登出
      description: |
        推荐的登出端点。浏览器会将 Path=/api/v1/oauth 的 refresh_token Cookie
        发送到该路径，后端可据此完整撤销 refresh token 与当前会话。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /roles:
    get:
      operationId: listRoles
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 自定义角色列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
    post:
      operationId: createRole
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '201':
          description: 角色已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      operationId: getRole
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 角色详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          $ref: '#/components/responses/Error'
    put:
      operationId: updateRole
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '200':
          description: 角色已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteRole
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Message'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /resource-grants:
    get:
      operationId: listResourceGrants
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: subject_type
          in: query
          schema:
            type: string
            enum: [user, role]
        - name: subject_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 资源授权列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGrantListResponse'
    post:
      operationId: grantResourcePermission
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceGrantRequest'
      responses:
        '201':
          description: 资源权限已授予
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGrant'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /resource-grants/revoke:
    post:
      operationId: revokeResourcePermission
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceGrantRequest'
      responses:
        '200':
          $ref: '#/components/responses/Message'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /servers:
    get:
      operationId: getServers
      tags:
        - servers
      summary: 获取服务器列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: group
          in: query
          description: 按组筛选
          schema:
            type: string
        - name: status
          in: query
          description: 按状态筛选
          schema:
            type: string
            enum: [online, offline, error]
      responses:
        '200':
          description: 成功获取服务器列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerList'

    post:
      operationId: postServers
      tags:
        - servers
      summary: 创建服务器
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '201':
          description: 服务器创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /servers/{id}:
    get:
      operationId: getServersId
      tags:
        - servers
      summary: 获取服务器详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          description: 服务器不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      operationId: putServersId
      tags:
        - servers
      summary: 更新服务器
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'

    delete:
      operationId: deleteServersId
      tags:
        - servers
      summary: 删除服务器
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功，无响应体

  # ========================================
  # SSH 会话（WebSocket 在实际使用时）
  # ========================================
...
  /scripts:
    get:
      operationId: getScripts
      tags:
        - scripts
      summary: 获取脚本列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Script'

    post:
      operationId: postScripts
      tags:
        - scripts
      summary: 创建脚本
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - content
              properties:
                name:
                  type: string
                description:
                  type: string
                content:
                  type: string
                language:
                  type: string
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Script'
  /scripts/{id}:
    get:
      operationId: getScriptsId
      tags:
        - scripts
      summary: 获取脚本详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Script'

    put:
      operationId: putScriptsId
      tags:
        - scripts
      summary: 更新脚本
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                content:
                  type: string
                language:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Script'

    delete:
      operationId: deleteScriptsId
      tags:
        - scripts
      summary: 删除脚本
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ========================================
  # 系统监控
  # ========================================
  # ========================================
  # 日志管理
  # ========================================
...
  /users:
    get:
      operationId: getUsers
      tags:
        - users
      summary: 获取用户列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'


    post:
      operationId: postUsers
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /users/me:
    get:
      operationId: getUsersMe
      tags:
        - users
      summary: 获取当前用户信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'


    put:
      operationId: putUsersMe
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /users/me/notifications:
    put:
      operationId: putUsersMeNotifications
      tags: [users, notifications]
      summary: 更新当前用户的通知偏好
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserNotificationSettingsRequest'
      responses:
        '200':
          description: 通知偏好已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string

  # ========================================
  # 统一备份 3.0（age）
  # ========================================
...
  /logs/cleanup:
    delete:
      operationId: deleteLogsCleanup
      tags: [logs]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
...
  /scheduled-tasks:
    get:
      operationId: getScheduledTasks
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postScheduledTasks
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /scheduled-tasks/{id}:
    get:
      operationId: getScheduledTasksId
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    put:
      operationId: putScheduledTasksId
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    delete:
      operationId: deleteScheduledTasksId
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /scheduled-tasks/{id}/toggle:
    post:
      operationId: postScheduledTasksIdToggle
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /scheduled-tasks/{id}/trigger:
    post:
      operationId: postScheduledTasksIdTrigger
      tags: [scheduled-tasks]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          description: 任务已进入持久化队列
          content:
            application/json:
              schema:
                type: object
                required: [message, task_run_id]
                properties:
                  message:
                    type: string
                  task_run_id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/Error'
...
  /scripts/{id}/execute:
    post:
      operationId: postScriptsIdExecute
      tags: [scripts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /servers/reorder:
    patch:
      operationId: patchServersReorder
      tags: [servers]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
...
  /settings/access-control:
    get:
      operationId: getSettingsAccessControl
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsAccessControl
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /settings/login-session:
    get:
      operationId: getSettingsLoginSession
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsLoginSession
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /settings/cors:
    get:
      operationId: getSettingsCors
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsCors
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /settings/login-security:
    get:
      operationId: getSettingsLoginSecurity
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsLoginSecurity
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /settings/ai/system:
    get:
      operationId: getSettingsAiSystem
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsAiSystem
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
...
  /settings/dingtalk:
    get:
      operationId: getSettingsDingtalk
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsDingtalk
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
...
  /settings/notifications:
    get:
      operationId: getSettingsNotifications
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsNotifications
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /settings/web-security:
    get:
      operationId: getSettingsWebSecurity
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsWebSecurity
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
  /settings/smtp:
    get:
      operationId: getSettingsSmtp
      tags: [settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'

    post:
      operationId: postSettingsSmtp
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
...
  /settings/system/basic:
    patch:
      operationId: patchSettingsSystemBasic
      tags: [settings]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONObject'
      responses:
        '200':
          $ref: '#/components/responses/JSON'
        '400':
          $ref: '#/components/responses/Error'
...